backend/requirements.txt
# Solidity to Clarity Converter - Requirements
Flask==2.2.3
flask-cors==3.0.10
requests==2.28.2
python-dotenv==1.0.0
gunicorn==20.1.0
